Separate RatingStar to a new file
